pyproject.toml
[project]
name = "soreq"
version = "0.1.0"
description = "A Python package for Soreq, a tool for managing and analyzing data."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "black>=25.1.0",
    "faiss-cpu>=1.11.0",
    "langchain>=0.3.25",
    "langchain-community>=0.3.24",
    "langchain-ollama>=0.3.3",
    "python-dotenv>=1.1.0",
    "streamlit>=1.45.1",
    "unstructured[md]>=0.17.2",
]
